+++ packages/components/src/components.d.ts
@@ -5,14 +5,14 @@
  * It contains typing information for all components that exist in this project.
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
-import { BreakpointCustomizable, ButtonAriaAttribute, ButtonType, ButtonVariant, IconName, LinkAriaAttribute, LinkTarget, LinkVariant, SelectedAriaAttributes, TextColor, TextSize, TextWeight, Theme } from "./types";
+import { BreakpointCustomizable, ButtonAriaAttribute, ButtonType, ButtonVariant, HeadingSize, IconName, LinkAriaAttribute, LinkTarget, LinkVariant, SelectedAriaAttributes, TextColor, TextSize, TextWeight, Theme } from "./types";
 import { AccordionSize, AccordionTag, AccordionUpdateEvent } from "./components/accordion/accordion-utils";
 import { BannerState, BannerWidth } from "./components/banner/banner-utils";
 import { ButtonIcon } from "./components/button/button-utils";
 import { ButtonGroupDirection } from "./components/button-group/button-group-utils";
 import { ButtonPureAlignLabel, ButtonPureAriaAttribute, ButtonPureIcon, ButtonPureSize, ButtonPureType, ButtonPureWeight } from "./components/button-pure/button-pure-utils";
 import { ButtonTileAlign, ButtonTileAriaAttribute, ButtonTileAspectRatio, ButtonTileBackground, ButtonTileIcon, ButtonTileSize, ButtonTileType, ButtonTileWeight } from "./components/button-tile/button-tile-utils";
-import { CarouselAlignHeader, CarouselInternationalization, CarouselUpdateEvent, CarouselWidth } from "./components/carousel/carousel-utils";
+import { CarouselAlignHeader, CarouselAriaAttribute, CarouselHeadingSize, CarouselInternationalization, CarouselUpdateEvent, CarouselWidth } from "./components/carousel/carousel-utils";
 import { CheckboxWrapperState } from "./components/checkbox-wrapper/checkbox-wrapper-utils";
 import { ContentWrapperBackgroundColor, ContentWrapperWidth } from "./components/content-wrapper/content-wrapper-utils";
 import { CrestAriaAttribute, CrestTarget } from "./components/crest/crest-utils";
@@ -23,10 +23,11 @@ import { FieldsetWrapperLabelSize, FieldsetWrapperState } from "./components/fie
 import { FlexAlignContent, FlexAlignItems, FlexDirection, FlexInline, FlexJustifyContent, FlexWrap } from "./components/flex/flex/flex-utils";
 import { FlexItemAlignSelf, FlexItemFlex, FlexItemGrow, FlexItemOffset, FlexItemShrink, FlexItemWidth } from "./components/flex/flex-item/flex-item-utils";
 import { FlyoutAriaAttribute, FlyoutPosition } from "./components/flyout/flyout-utils";
+import { FlyoutNavigationAriaAttribute, FlyoutNavigationUpdateEvent } from "./components/flyout-navigation/flyout-navigation/flyout-navigation-utils";
 import { GridDirection, GridGutter, GridWrap } from "./components/grid/grid/grid-utils";
 import { GridItemOffset, GridItemSize } from "./components/grid/grid-item/grid-item-utils";
 import { HeadingTag } from "./components/heading/heading-tag";
-import { HeadingAlign, HeadingColor, HeadingSize } from "./components/heading/heading-utils";
+import { HeadingAlign, HeadingColor } from "./components/heading/heading-utils";
 import { HeadlineAlign, HeadlineColor, HeadlineTag, HeadlineVariant } from "./components/headline/headline-utils";
 import { IconAriaAttribute, IconColor, IconSize } from "./components/icon/icon-utils";
 import { InlineNotificationActionIcon, InlineNotificationState } from "./components/inline-notification/inline-notification-utils";
@@ -35,6 +36,7 @@ import { LinkPureAlignLabel, LinkPureAriaAttribute, LinkPureIcon, LinkPureSize, 
 import { LinkSocialIcon, LinkSocialTarget } from "./components/link-social/link-social-utils";
 import { LinkTileAlign, LinkTileAriaAttribute, LinkTileAspectRatio, LinkTileBackground, LinkTileSize, LinkTileTarget, LinkTileWeight } from "./components/link-tile/link-tile-utils";
 import { LinkTileModelSignatureAspectRatio, LinkTileModelSignatureHeadingTag, LinkTileModelSignatureLinkDirection, LinkTileModelSignatureModel, LinkTileModelSignatureWeight } from "./components/link-tile-model-signature/link-tile-model-signature-utils";
+import { LinkTileProductAspectRatio, LinkTileProductLikeEvent, LinkTileProductTarget } from "./components/link-tile-product/link-tile-product-utils";
 import { MarqueAriaAttribute, MarqueTarget, MarqueVariant } from "./components/marque/marque-utils";
 import { MarqueSize } from "./components/marque/marque-size";
 import { ModalAriaAttribute } from "./components/modal/modal-utils";
@@ -64,14 +66,14 @@ import { TextareaWrapperState } from "./components/textarea-wrapper/textarea-wra
 import { ToastMessage } from "./components/toast/toast/toast-manager";
 import { ToastState } from "./components/toast/toast/toast-utils";
 import { WordmarkAriaAttribute, WordmarkSize, WordmarkTarget } from "./components/wordmark/wordmark-utils";
-export { BreakpointCustomizable, ButtonAriaAttribute, ButtonType, ButtonVariant, IconName, LinkAriaAttribute, LinkTarget, LinkVariant, SelectedAriaAttributes, TextColor, TextSize, TextWeight, Theme } from "./types";
+export { BreakpointCustomizable, ButtonAriaAttribute, ButtonType, ButtonVariant, HeadingSize, IconName, LinkAriaAttribute, LinkTarget, LinkVariant, SelectedAriaAttributes, TextColor, TextSize, TextWeight, Theme } from "./types";
 export { AccordionSize, AccordionTag, AccordionUpdateEvent } from "./components/accordion/accordion-utils";
 export { BannerState, BannerWidth } from "./components/banner/banner-utils";
 export { ButtonIcon } from "./components/button/button-utils";
 export { ButtonGroupDirection } from "./components/button-group/button-group-utils";
 export { ButtonPureAlignLabel, ButtonPureAriaAttribute, ButtonPureIcon, ButtonPureSize, ButtonPureType, ButtonPureWeight } from "./components/button-pure/button-pure-utils";
 export { ButtonTileAlign, ButtonTileAriaAttribute, ButtonTileAspectRatio, ButtonTileBackground, ButtonTileIcon, ButtonTileSize, ButtonTileType, ButtonTileWeight } from "./components/button-tile/button-tile-utils";
-export { CarouselAlignHeader, CarouselInternationalization, CarouselUpdateEvent, CarouselWidth } from "./components/carousel/carousel-utils";
+export { CarouselAlignHeader, CarouselAriaAttribute, CarouselHeadingSize, CarouselInternationalization, CarouselUpdateEvent, CarouselWidth } from "./components/carousel/carousel-utils";
 export { CheckboxWrapperState } from "./components/checkbox-wrapper/checkbox-wrapper-utils";
 export { ContentWrapperBackgroundColor, ContentWrapperWidth } from "./components/content-wrapper/content-wrapper-utils";
 export { CrestAriaAttribute, CrestTarget } from "./components/crest/crest-utils";
@@ -82,10 +84,11 @@ export { FieldsetWrapperLabelSize, FieldsetWrapperState } from "./components/fie
 export { FlexAlignContent, FlexAlignItems, FlexDirection, FlexInline, FlexJustifyContent, FlexWrap } from "./components/flex/flex/flex-utils";
 export { FlexItemAlignSelf, FlexItemFlex, FlexItemGrow, FlexItemOffset, FlexItemShrink, FlexItemWidth } from "./components/flex/flex-item/flex-item-utils";
 export { FlyoutAriaAttribute, FlyoutPosition } from "./components/flyout/flyout-utils";
+export { FlyoutNavigationAriaAttribute, FlyoutNavigationUpdateEvent } from "./components/flyout-navigation/flyout-navigation/flyout-navigation-utils";
 export { GridDirection, GridGutter, GridWrap } from "./components/grid/grid/grid-utils";
 export { GridItemOffset, GridItemSize } from "./components/grid/grid-item/grid-item-utils";
 export { HeadingTag } from "./components/heading/heading-tag";
-export { HeadingAlign, HeadingColor, HeadingSize } from "./components/heading/heading-utils";
+export { HeadingAlign, HeadingColor } from "./components/heading/heading-utils";
 export { HeadlineAlign, HeadlineColor, HeadlineTag, HeadlineVariant } from "./components/headline/headline-utils";
 export { IconAriaAttribute, IconColor, IconSize } from "./components/icon/icon-utils";
 export { InlineNotificationActionIcon, InlineNotificationState } from "./components/inline-notification/inline-notification-utils";
@@ -94,6 +97,7 @@ export { LinkPureAlignLabel, LinkPureAriaAttribute, LinkPureIcon, LinkPureSize, 
 export { LinkSocialIcon, LinkSocialTarget } from "./components/link-social/link-social-utils";
 export { LinkTileAlign, LinkTileAriaAttribute, LinkTileAspectRatio, LinkTileBackground, LinkTileSize, LinkTileTarget, LinkTileWeight } from "./components/link-tile/link-tile-utils";
 export { LinkTileModelSignatureAspectRatio, LinkTileModelSignatureHeadingTag, LinkTileModelSignatureLinkDirection, LinkTileModelSignatureModel, LinkTileModelSignatureWeight } from "./components/link-tile-model-signature/link-tile-model-signature-utils";
+export { LinkTileProductAspectRatio, LinkTileProductLikeEvent, LinkTileProductTarget } from "./components/link-tile-product/link-tile-product-utils";
 export { MarqueAriaAttribute, MarqueTarget, MarqueVariant } from "./components/marque/marque-utils";
 export { MarqueSize } from "./components/marque/marque-size";
 export { ModalAriaAttribute } from "./components/modal/modal-utils";
@@ -372,6 +376,10 @@ export namespace Components {
          */
         "alignHeader"?: CarouselAlignHeader;
         /**
+          * Add ARIA attributes.
+         */
+        "aria"?: SelectedAriaAttributes<CarouselAriaAttribute>;
+        /**
           * Defines the description used in the carousel.
          */
         "description"?: string;
@@ -383,6 +391,10 @@ export namespace Components {
           * Defines the heading used in the carousel.
          */
         "heading"?: string;
+        /**
+          * Defines the heading size used in the carousel.
+         */
+        "headingSize"?: CarouselHeadingSize;
         /**
           * Override the default wordings that are used for aria-labels on the next/prev buttons and pagination.
          */
@@ -650,6 +662,31 @@ export namespace Components {
           * Adapts the flyout color depending on the theme.
          */
         "theme"?: Theme;
+    }
+    /**
+     * __Experimental__
+     */
+    interface PFlyoutNavigation {
+        /**
+          * Defines which flyout-navigation-item to be visualized as opened.
+         */
+        "activeIdentifier"?: string | undefined;
+        /**
+          * Add ARIA attributes.
+         */
+        "aria"?: SelectedAriaAttributes<FlyoutNavigationAriaAttribute>;
+        /**
+          * If true, the flyout-navigation is visualized as opened.
+         */
+        "open"?: boolean;
+        /**
+          * Adapts the flyout-navigation color depending on the theme.
+         */
+        "theme"?: Theme;
+    }
+    interface PFlyoutNavigationItem {
+        "identifier": string;
+        "label"?: string;
     }
     /**
      * @deprecated since v3.0.0, will be removed with next major release. Use native CSS Grid instead.
@@ -1031,6 +1068,51 @@ export namespace Components {
           * Adapts the font weight of the heading.
          */
         "weight"?: BreakpointCustomizable<LinkTileModelSignatureWeight>;
+    }
+    /**
+     * __Experimental__
+     */
+    interface PLinkTileProduct {
+        /**
+          * Aspect ratio of the link-tile-product.
+         */
+        "aspectRatio"?: BreakpointCustomizable<LinkTileProductAspectRatio>;
+        /**
+          * Additional product description
+         */
+        "description"?: string;
+        /**
+          * Product heading
+         */
+        "heading": string;
+        /**
+          * href of the `<a>`.
+         */
+        "href"?: string;
+        /**
+          * A Boolean attribute indicating that a like button should be shown.
+         */
+        "likeButton"?: boolean;
+        /**
+          * A Boolean attribute indicating that a product is liked.
+         */
+        "liked"?: boolean;
+        /**
+          * Product price
+         */
+        "price": string;
+        /**
+          * Specifies the relationship of the target object to the link object.
+         */
+        "rel"?: string;
+        /**
+          * Target attribute where the link should be opened.
+         */
+        "target"?: LinkTileProductTarget;
+        /**
+          * Adapts the link-tile-product color depending on the theme.
+         */
+        "theme"?: Theme;
     }
     /**
      * @deprecated since v3.0.0, will be removed with next major release. Please use "p-wordmark" instead.
@@ -1839,9 +1921,17 @@ export interface PFlyoutCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLPFlyoutElement;
 }
+export interface PFlyoutNavigationCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLPFlyoutNavigationElement;
+}
 export interface PInlineNotificationCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLPInlineNotificationElement;
+}
+export interface PLinkTileProductCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLPLinkTileProductElement;
 }
 export interface PModalCustomEvent<T> extends CustomEvent<T> {
     detail: T;
@@ -2052,6 +2142,33 @@ declare global {
         prototype: HTMLPFlyoutElement;
         new (): HTMLPFlyoutElement;
     };
+    interface HTMLPFlyoutNavigationElementEventMap {
+        "dismiss": void;
+        "update": FlyoutNavigationUpdateEvent;
+    }
+    /**
+     * __Experimental__
+     */
+    interface HTMLPFlyoutNavigationElement extends Components.PFlyoutNavigation, HTMLStencilElement {
+        addEventListener<K extends keyof HTMLPFlyoutNavigationElementEventMap>(type: K, listener: (this: HTMLPFlyoutNavigationElement, ev: PFlyoutNavigationCustomEvent<HTMLPFlyoutNavigationElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLPFlyoutNavigationElementEventMap>(type: K, listener: (this: HTMLPFlyoutNavigationElement, ev: PFlyoutNavigationCustomEvent<HTMLPFlyoutNavigationElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
+    }
+    var HTMLPFlyoutNavigationElement: {
+        prototype: HTMLPFlyoutNavigationElement;
+        new (): HTMLPFlyoutNavigationElement;
+    };
+    interface HTMLPFlyoutNavigationItemElement extends Components.PFlyoutNavigationItem, HTMLStencilElement {
+    }
+    var HTMLPFlyoutNavigationItemElement: {
+        prototype: HTMLPFlyoutNavigationItemElement;
+        new (): HTMLPFlyoutNavigationItemElement;
+    };
     /**
      * @deprecated since v3.0.0, will be removed with next major release. Use native CSS Grid instead.
      */
@@ -2141,6 +2258,26 @@ declare global {
     var HTMLPLinkTileModelSignatureElement: {
         prototype: HTMLPLinkTileModelSignatureElement;
         new (): HTMLPLinkTileModelSignatureElement;
+    };
+    interface HTMLPLinkTileProductElementEventMap {
+        "like": LinkTileProductLikeEvent;
+    }
+    /**
+     * __Experimental__
+     */
+    interface HTMLPLinkTileProductElement extends Components.PLinkTileProduct, HTMLStencilElement {
+        addEventListener<K extends keyof HTMLPLinkTileProductElementEventMap>(type: K, listener: (this: HTMLPLinkTileProductElement, ev: PLinkTileProductCustomEvent<HTMLPLinkTileProductElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLPLinkTileProductElementEventMap>(type: K, listener: (this: HTMLPLinkTileProductElement, ev: PLinkTileProductCustomEvent<HTMLPLinkTileProductElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
+    }
+    var HTMLPLinkTileProductElement: {
+        prototype: HTMLPLinkTileProductElement;
+        new (): HTMLPLinkTileProductElement;
     };
     /**
      * @deprecated since v3.0.0, will be removed with next major release. Please use "p-wordmark" instead.
@@ -2531,6 +2668,8 @@ declare global {
         "p-flex": HTMLPFlexElement;
         "p-flex-item": HTMLPFlexItemElement;
         "p-flyout": HTMLPFlyoutElement;
+        "p-flyout-navigation": HTMLPFlyoutNavigationElement;
+        "p-flyout-navigation-item": HTMLPFlyoutNavigationItemElement;
         "p-grid": HTMLPGridElement;
         "p-grid-item": HTMLPGridItemElement;
         "p-heading": HTMLPHeadingElement;
@@ -2542,6 +2681,7 @@ declare global {
         "p-link-social": HTMLPLinkSocialElement;
         "p-link-tile": HTMLPLinkTileElement;
         "p-link-tile-model-signature": HTMLPLinkTileModelSignatureElement;
+        "p-link-tile-product": HTMLPLinkTileProductElement;
         "p-marque": HTMLPMarqueElement;
         "p-modal": HTMLPModalElement;
         "p-model-signature": HTMLPModelSignatureElement;
@@ -2843,6 +2983,10 @@ declare namespace LocalJSX {
          */
         "alignHeader"?: CarouselAlignHeader;
         /**
+          * Add ARIA attributes.
+         */
+        "aria"?: SelectedAriaAttributes<CarouselAriaAttribute>;
+        /**
           * Defines the description used in the carousel.
          */
         "description"?: string;
@@ -2854,6 +2998,10 @@ declare namespace LocalJSX {
           * Defines the heading used in the carousel.
          */
         "heading"?: string;
+        /**
+          * Defines the heading size used in the carousel.
+         */
+        "headingSize"?: CarouselHeadingSize;
         /**
           * Override the default wordings that are used for aria-labels on the next/prev buttons and pagination.
          */
@@ -3133,6 +3281,39 @@ declare namespace LocalJSX {
           * Adapts the flyout color depending on the theme.
          */
         "theme"?: Theme;
+    }
+    /**
+     * __Experimental__
+     */
+    interface PFlyoutNavigation {
+        /**
+          * Defines which flyout-navigation-item to be visualized as opened.
+         */
+        "activeIdentifier"?: string | undefined;
+        /**
+          * Add ARIA attributes.
+         */
+        "aria"?: SelectedAriaAttributes<FlyoutNavigationAriaAttribute>;
+        /**
+          * Emitted when the component requests to be dismissed.
+         */
+        "onDismiss"?: (event: PFlyoutNavigationCustomEvent<void>) => void;
+        /**
+          * Emitted when activeIdentifier is changed.
+         */
+        "onUpdate"?: (event: PFlyoutNavigationCustomEvent<FlyoutNavigationUpdateEvent>) => void;
+        /**
+          * If true, the flyout-navigation is visualized as opened.
+         */
+        "open"?: boolean;
+        /**
+          * Adapts the flyout-navigation color depending on the theme.
+         */
+        "theme"?: Theme;
+    }
+    interface PFlyoutNavigationItem {
+        "identifier"?: string;
+        "label"?: string;
     }
     /**
      * @deprecated since v3.0.0, will be removed with next major release. Use native CSS Grid instead.
@@ -3522,6 +3703,55 @@ declare namespace LocalJSX {
           * Adapts the font weight of the heading.
          */
         "weight"?: BreakpointCustomizable<LinkTileModelSignatureWeight>;
+    }
+    /**
+     * __Experimental__
+     */
+    interface PLinkTileProduct {
+        /**
+          * Aspect ratio of the link-tile-product.
+         */
+        "aspectRatio"?: BreakpointCustomizable<LinkTileProductAspectRatio>;
+        /**
+          * Additional product description
+         */
+        "description"?: string;
+        /**
+          * Product heading
+         */
+        "heading"?: string;
+        /**
+          * href of the `<a>`.
+         */
+        "href"?: string;
+        /**
+          * A Boolean attribute indicating that a like button should be shown.
+         */
+        "likeButton"?: boolean;
+        /**
+          * A Boolean attribute indicating that a product is liked.
+         */
+        "liked"?: boolean;
+        /**
+          * Emitted when the like button is clicked.
+         */
+        "onLike"?: (event: PLinkTileProductCustomEvent<LinkTileProductLikeEvent>) => void;
+        /**
+          * Product price
+         */
+        "price"?: string;
+        /**
+          * Specifies the relationship of the target object to the link object.
+         */
+        "rel"?: string;
+        /**
+          * Target attribute where the link should be opened.
+         */
+        "target"?: LinkTileProductTarget;
+        /**
+          * Adapts the link-tile-product color depending on the theme.
+         */
+        "theme"?: Theme;
     }
     /**
      * @deprecated since v3.0.0, will be removed with next major release. Please use "p-wordmark" instead.
@@ -4410,6 +4640,8 @@ declare namespace LocalJSX {
         "p-flex": PFlex;
         "p-flex-item": PFlexItem;
         "p-flyout": PFlyout;
+        "p-flyout-navigation": PFlyoutNavigation;
+        "p-flyout-navigation-item": PFlyoutNavigationItem;
         "p-grid": PGrid;
         "p-grid-item": PGridItem;
         "p-heading": PHeading;
@@ -4421,6 +4653,7 @@ declare namespace LocalJSX {
         "p-link-social": PLinkSocial;
         "p-link-tile": PLinkTile;
         "p-link-tile-model-signature": PLinkTileModelSignature;
+        "p-link-tile-product": PLinkTileProduct;
         "p-marque": PMarque;
         "p-modal": PModal;
         "p-model-signature": PModelSignature;
@@ -4495,6 +4728,11 @@ declare module "@stencil/core" {
             "p-flex-item": LocalJSX.PFlexItem & JSXBase.HTMLAttributes<HTMLPFlexItemElement>;
             "p-flyout": LocalJSX.PFlyout & JSXBase.HTMLAttributes<HTMLPFlyoutElement>;
             /**
+             * __Experimental__
+             */
+            "p-flyout-navigation": LocalJSX.PFlyoutNavigation & JSXBase.HTMLAttributes<HTMLPFlyoutNavigationElement>;
+            "p-flyout-navigation-item": LocalJSX.PFlyoutNavigationItem & JSXBase.HTMLAttributes<HTMLPFlyoutNavigationItemElement>;
+            /**
              * @deprecated since v3.0.0, will be removed with next major release. Use native CSS Grid instead.
              */
             "p-grid": LocalJSX.PGrid & JSXBase.HTMLAttributes<HTMLPGridElement>;
@@ -4517,6 +4755,10 @@ declare module "@stencil/core" {
             "p-link-social": LocalJSX.PLinkSocial & JSXBase.HTMLAttributes<HTMLPLinkSocialElement>;
             "p-link-tile": LocalJSX.PLinkTile & JSXBase.HTMLAttributes<HTMLPLinkTileElement>;
             "p-link-tile-model-signature": LocalJSX.PLinkTileModelSignature & JSXBase.HTMLAttributes<HTMLPLinkTileModelSignatureElement>;
+            /**
+             * __Experimental__
+             */
+            "p-link-tile-product": LocalJSX.PLinkTileProduct & JSXBase.HTMLAttributes<HTMLPLinkTileProductElement>;
             /**
              * @deprecated since v3.0.0, will be removed with next major release. Please use "p-wordmark" instead.
              */