+++ packages/components/src/components/flyout-navigation/flyout-navigation-item/flyout-navigation-item-utils.ts
@@ -1,0 +1,6 @@
+import { type Theme } from '../../../utils';
+
+export type FlyoutNavigationItemInternalHTMLProps = {
+  open: boolean;
+  theme: Theme;
+};