+++ packages/components/src/utils/dom/index.ts
@@ -1,5 +1,6 @@
 export * from './dom-types';
 export * from './getAttribute';
+export * from './getDirectChildHTMLElementOfKind';
 export * from './getClosestHTMLElement';
 export * from './getDirectChildHTMLElement';
 export * from './getDirectChildHTMLElements';