+++ packages/storefront/src/components/CodeEditor.vue
@@ -14,7 +14,7 @@
   import Vue from 'vue';
   import Component from 'vue-class-component';
   import { Prop } from 'vue-property-decorator';
-  import type { BackgroundColor, Framework, PlaygroundTheme } from '@/models';
+  import type { BackgroundColor, Framework, PlaygroundDir, PlaygroundTheme } from '@/models';
   import { openInStackBlitz } from '@/utils';
   import type { ExternalDependency, SharedImportKey } from '@/utils';
   import { isStableStorefrontRelease } from '@/utils/stackblitz/helper';
@@ -26,6 +26,7 @@
   export default class CodeEditor extends Vue {
     @Prop({ default: '' }) public markup!: string;
     @Prop({ default: 'light' }) public theme!: PlaygroundTheme;
+    @Prop({ default: 'ltr' }) public dir!: PlaygroundDir;
     @Prop({ default: 'vanilla-js' }) public framework!: Exclude<Framework, 'shared' | 'vue'>; // we don't have stackblitz integration for vue yet, therefore excluding vue
     @Prop({ default: 'background-base' }) public backgroundColor!: BackgroundColor;
     @Prop({ default: () => [] }) public externalStackBlitzDependencies!: ExternalDependency[];
@@ -43,6 +44,7 @@
         markup: this.markup,
         framework: this.framework,
         theme: this.theme,
+        dir: this.dir,
         externalDependencies: this.externalStackBlitzDependencies,
         backgroundColor: this.backgroundColor,
         sharedImportKeys: this.sharedImportKeys,