+++ packages/components/CHANGELOG.md
@@ -16,6 +16,36 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### Added
 
+- `Carousel`: Prop `heading-size`, named slot `header`
+  ([#2915](https://github.com/porsche-design-system/porsche-design-system/pull/2915))
+
+#### Changed
+
+- `Carousel`: Position and width of heading and description
+  ([#2915](https://github.com/porsche-design-system/porsche-design-system/pull/2915))
+
+### [3.10.0-rc.3] - 2023-12-12
+
+### [3.10.0-rc.2] - 2023-12-12
+
+### [3.10.0-rc.1] - 2023-12-11
+
+#### Added
+
+- **[EXPERIMENTAL]** `Link Tile Product`
+  ([#2909](https://github.com/porsche-design-system/porsche-design-system/pull/2909))
+
+#### Fixed
+
+- `Wordmark`, `Crest` and `Marque`: custom clickable area
+  ([#2930](https://github.com/porsche-design-system/porsche-design-system/pull/2930))
+
+### [3.10.0-rc.0] - 2023-12-07
+
+#### Added
+
+- **[EXPERIMENTAL]** `Flyout Navigation`
+  ([#2906](https://github.com/porsche-design-system/porsche-design-system/pull/2906))
 - Prop `submit-button` to show/hide a submit button for `Text Field Wrapper` `type="search"` if wrapped inside a form
   ([#2908](https://github.com/porsche-design-system/porsche-design-system/pull/2908))
 
@@ -23,12 +53,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - `Accordion`: removed `border-bottom` if used standalone
   ([#2911](https://github.com/porsche-design-system/porsche-design-system/pull/2911))
-- `display` css property can be overridden on all components, e.g. within media queries  
+- `display` css property can be overridden on all components, e.g. to make use of `display: none;` within media
+  queries  
   ([#2913](https://github.com/porsche-design-system/porsche-design-system/pull/2913))
 - `Pagination`: Prop `maxNumberOfPageLinks` is deprecated and has no effect anymore, instead there is responsive
   behavior out of the box with full SSR support
   ([#2898](https://github.com/porsche-design-system/porsche-design-system/pull/2898))
-
 
 ### [3.9.0] - 2023-11-24